pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "clinical-trial-env"
version = "0.2.0"
description = "Clinical Trial Optimization RL Environment — 5 tasks, 10 actions, multi-organ safety"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openenv-core>=0.2.1",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "openai>=1.20.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
server = "server.app:main"
